requirementsDB.txt

feedparser
Jinja2
PyYAML
python-dateutil
python-dotenv
requests
openai
jieba
duckdb
